fix(compiler): JIT unit tests: Don't compare expressions of different signedness
Fix compiler warnings in `end_to_end_jit_clear_tensor.cc` and
`end_to_end_jit_encrypted_tensor.cc` due to the comparison of integer
expressions of different signedness.

+++ compiler/tests/unittest/end_to_end_jit_clear_tensor.cc
@@ -222,8 +222,8 @@ func @main(%t: tensor<2x10xi64>, %i: index, %j: index) -> i64 {
       mlir::zamalang::IntLambdaArgument<uint64_t>>
       arg(tensor2D, shape2D);
 
-  for (size_t i = 0; i < dims[0]; i++) {
-    for (size_t j = 0; j < dims[1]; j++) {
+  for (int64_t i = 0; i < dims[0]; i++) {
+    for (int64_t j = 0; j < dims[1]; j++) {
       auto pos = i * dims[1] + j;
       mlir::zamalang::IntLambdaArgument<size_t> argi(i);
       mlir::zamalang::IntLambdaArgument<size_t> argj(j);
@@ -251,7 +251,7 @@ func @main(%t: tensor<2x10xi64>) -> tensor<1x5xi64> {
 
   ASSERT_EXPECTED_SUCCESS(res);
 
-  ASSERT_EQ(res->size(), 1 * 5);
+  ASSERT_EQ(res->size(), (size_t)1 * 5);
 
   // Check the sub slice
   for (size_t j = 0; j < 5; j++) {
@@ -279,7 +279,7 @@ func @main(%t: tensor<2x10xi64>) -> tensor<1x5xi64> {
 
   ASSERT_EXPECTED_SUCCESS(res);
 
-  ASSERT_EQ(res->size(), 1 * 5);
+  ASSERT_EQ(res->size(), (size_t)1 * 5);
 
   // Check the sub slice
   for (size_t j = 0; j < 5; j++) {
